To handle the type substitution correctly, we need JAX-WS 2.1.
Otherwise, this problem manifests itself as a mysterious disappearance
of some fields when you send a request --- such as TraversalSpec
marshalled as its base class "SelectionSpec".

This is because JAX-WS at runtime doesn't notice the presence of
the TraversalSpec class, as such a class is not reachable from the
SEI. JAX-WS 2.1 fixes this problem by introducing a necessary
annotation on SEI.

The only environment that has JAX-WS 2.0 in it is JDK6 initial release
up to JDK6u4 or something, so the dependency to 2.1 shouldn't cause
a problem.